+++ merged-packages/bitcoin-wallet-snap/src/infra/index.ts
@@ -1,4 +1,5 @@
 export * from './BdkAccountAdapter';
+export * from './StoredAccountAdapter';
 export * from './SnapClientAdapter';
 export * from './EsploraClientAdapter';
 export * from './PriceApiClientAdapter';